Removed duplicate classes and tests after rebase
