+++ tests/devices/unit_tests/electron_analyser/test_specs_region.py
@@ -1,16 +1,27 @@
-import os
+from typing import Any
 
-from tests.devices.unit_tests.electron_analyser.utils import (
-    TEST_DATA_PATH,
+import pytest
+from tests.devices.unit_tests.electron_analyser.test_utils import (
     check_region_model_list_to_expected_values,
+    is_list_of_custom_type,
 )
 
-from dodal.common.data_util import load_json_file_to_class
-from dodal.devices.electron_analyser.base_region import EnergyMode
-from dodal.devices.electron_analyser.specs.specs_region import SpecsSequence
+from dodal.devices.electron_analyser.abstract_region import EnergyMode
+from dodal.devices.electron_analyser.specs_region import SpecsRegion, SpecsSequence
 
 
-def get_expected_region_values():
+@pytest.fixture
+def sequence_file() -> str:
+    return "specs_sequence.seq"
+
+
+@pytest.fixture
+def sequence_class() -> type[SpecsSequence]:
+    return SpecsSequence
+
+
+@pytest.fixture
+def expected_region_values() -> list[dict[str, Any]]:
     return [
         {
             "name": "region",
@@ -51,13 +62,51 @@ def get_expected_region_values():
     ]
 
 
-def test_specs_file_loads_into_class():
-    file = os.path.join(TEST_DATA_PATH, "specs_sequence.seq")
-    sequence = load_json_file_to_class(SpecsSequence, file)
+@pytest.fixture
+def expected_region_names(
+    expected_region_values: list[dict[str, Any]],
+) -> list[str]:
+    names = []
+    for expected_region_value in expected_region_values:
+        names.append(expected_region_value["name"])
+    return names
 
-    assert sequence.get_region_names() == ["region", "region2"]
+
+def test_get_expected_region_from_name(
+    sequence: SpecsSequence, expected_region_names: list[str]
+) -> None:
+    for name in expected_region_names:
+        assert sequence.get_region_by_name(name) is not None
+    assert sequence.get_region_by_name("region name should not be in sequence") is None
+
+
+def test_sequence_get_expected_region_type(sequence: SpecsSequence) -> None:
+    assert is_list_of_custom_type(sequence.regions, SpecsRegion)
+    assert is_list_of_custom_type(sequence.get_enabled_regions(), SpecsRegion)
+
+
+def test_sequence_get_expected_region_names(
+    sequence: SpecsSequence, expected_region_names: list[str]
+) -> None:
+    assert sequence.get_region_names() == expected_region_names
+
+
+def test_sequence_get_expected_enabled_region_names(sequence: SpecsSequence) -> None:
     assert sequence.get_enabled_region_names() == ["region2"]
 
-    check_region_model_list_to_expected_values(
-        sequence.regions, get_expected_region_values()
-    )
+
+def test_region_kinetic_and_binding_energy(sequence: SpecsSequence) -> None:
+    for r in sequence.regions:
+        is_binding_energy = r.energyMode == EnergyMode.BINDING
+        is_kinetic_energy = r.energyMode == EnergyMode.KINETIC
+
+        assert r.is_binding_energy() == is_binding_energy
+        assert r.is_binding_energy() != is_kinetic_energy
+        assert r.is_kinetic_energy() == is_kinetic_energy
+        assert r.is_kinetic_energy() != is_binding_energy
+
+
+def test_specs_file_loads_into_class(
+    sequence: SpecsSequence, expected_region_values: list[dict[str, Any]]
+) -> None:
+    check_region_model_list_to_expected_values(sequence.regions, expected_region_values)
