+++ tests/devices/unit_tests/electron_analyser/test_vgscienta_region.py
@@ -1,21 +1,33 @@
-import os
+from typing import Any
 
-from tests.devices.unit_tests.electron_analyser.utils import (
-    TEST_DATA_PATH,
+import pytest
+from tests.devices.unit_tests.electron_analyser.test_utils import (
     check_region_model_list_to_expected_values,
+    is_list_of_custom_type,
 )
 
-from dodal.common.data_util import load_json_file_to_class
-from dodal.devices.electron_analyser.base_region import EnergyMode
-from dodal.devices.electron_analyser.vgscienta.vgscienta_region import (
+from dodal.devices.electron_analyser.abstract_region import EnergyMode
+from dodal.devices.electron_analyser.vgscienta_region import (
     AcquisitionMode,
     DetectorMode,
     Status,
+    VGScientaRegion,
     VGScientaSequence,
 )
 
 
-def get_expected_region_values():
+@pytest.fixture
+def sequence_file() -> str:
+    return "vgscienta_sequence.seq"
+
+
+@pytest.fixture
+def sequence_class() -> type[VGScientaSequence]:
+    return VGScientaSequence
+
+
+@pytest.fixture
+def expected_region_values() -> list[dict[str, Any]]:
     return [
         {
             "name": "New_Region",
@@ -72,13 +84,44 @@ def get_expected_region_values():
     ]
 
 
-def test_vgscienta_file_loads_into_class():
-    file = os.path.join(TEST_DATA_PATH, "vgscienta_sequence.seq")
-    sequence = load_json_file_to_class(VGScientaSequence, file)
+@pytest.fixture
+def expected_region_names(
+    expected_region_values: list[dict[str, Any]],
+) -> list[str]:
+    names = []
+    for expected_region_value in expected_region_values:
+        names.append(expected_region_value["name"])
+    return names
 
-    assert sequence.get_region_names() == ["New_Region", "New_Region1"]
+
+def test_get_expected_region_from_name(
+    sequence: VGScientaSequence, expected_region_names: list[str]
+) -> None:
+    for name in expected_region_names:
+        assert sequence.get_region_by_name(name) is not None
+    assert sequence.get_region_by_name("region name should not be in sequence") is None
+
+
+def test_sequence_get_expected_region_names(
+    sequence: VGScientaSequence, expected_region_names: list[str]
+) -> None:
+    assert sequence.get_region_names() == expected_region_names
+
+
+def test_sequence_get_expected_enabled_region_names(
+    sequence: VGScientaSequence,
+) -> None:
     assert sequence.get_enabled_region_names() == ["New_Region"]
 
+
+def test_sequence_get_expected_region_type(sequence: VGScientaSequence) -> None:
+    assert is_list_of_custom_type(sequence.regions, VGScientaRegion)
+    assert is_list_of_custom_type(sequence.get_enabled_regions(), VGScientaRegion)
+
+
+def test_sequence_get_expected_excitation_energy_source(
+    sequence: VGScientaSequence,
+) -> None:
     assert (
         sequence.get_excitation_energy_source_by_region(sequence.regions[0])
         == sequence.excitationEnergySources[1]
@@ -87,6 +130,20 @@ def test_vgscienta_file_loads_into_class():
         sequence.get_excitation_energy_source_by_region(sequence.regions[1])
         == sequence.excitationEnergySources[0]
     )
-    check_region_model_list_to_expected_values(
-        sequence.regions, get_expected_region_values()
-    )
+
+
+def test_region_kinetic_and_binding_energy(sequence: VGScientaSequence) -> None:
+    for r in sequence.regions:
+        is_binding_energy = r.energyMode == EnergyMode.BINDING
+        is_kinetic_energy = r.energyMode == EnergyMode.KINETIC
+
+        assert r.is_binding_energy() == is_binding_energy
+        assert r.is_binding_energy() != is_kinetic_energy
+        assert r.is_kinetic_energy() == is_kinetic_energy
+        assert r.is_kinetic_energy() != is_binding_energy
+
+
+def test_sequence_file_loads_into_sequence_class(
+    sequence: VGScientaSequence, expected_region_values: list[dict[str, Any]]
+) -> None:
+    check_region_model_list_to_expected_values(sequence.regions, expected_region_values)
